updated search and database pword uname consts
Search needed updated to be prevent showing a user its own tools,
tools that are lent and tools that have their status set to pending
after a user requests to borrow the tool. This commit also updates
the tools status after a request has been rejected from pending to
notLent so it can be searched again. Also removed pg password and
username and set it to ENV vars.