Added a map which shows location of requester and updated to httpCLient Module
Updated all the angular files associated with search from using httpmodule
to httpClientModule. This was needed because when trying to call APIs
with httpModule caused a X-CSRF-Token header is not allowed in pre request flight.
Once the http module was changed, an http call to google geocoding api was created,
which then parses the return for the lat and lng and passes it to the template
which AGM uses them to render the map.