Added request's mailer, fixed FK violation bug, cleaned up ctlr/angular code
Added mailer for when requests are created to be sent to the requester
and include info about about the requestee and tool. There was an FK
violation bug when attempting to delete a user, which was fixed
by added dependent: :destory in the users model and adding user_id
as a FK to requests table. Finally, fixed ruby controller code to
render new angular http module properly.