Change sidebard UX and table's UX
Changed the side by from a lists of links to a solid color sidbar that
exetends from header to footer and shows the links. The color and style
were change so the bar is purple and the links when hovered change from
white text to purple text over white background. Changed the UX of tables
throughout the app to be an ol of li's. In each li new fields were added
like tool status and a change css effect on hover.